Move RunLogout to kclient
