+++ pkg/odo/cli/logout/logout.go
@@ -41,7 +41,7 @@ func (o *LogoutOptions) Validate() (err error) {
 
 // Run contains the logic for the odo logout command
 func (o *LogoutOptions) Run(cmd *cobra.Command) (err error) {
-	return o.Client.RunLogout(os.Stdout)
+	return o.KClient.RunLogout(os.Stdout)
 }
 
 // NewCmdLogout implements the logout odo command
